Many, many improvements to chat

+++ app/controllers/messages_controller.rb
@@ -10,10 +10,10 @@ class MessagesController < ApplicationController
       .order(created_at: :desc)
       .limit(20)
     if params[:after]
-      @messages = @messages.created_after(Time.at(params[:after].to_i))
+      @messages = @messages.created_after(Time.at(params[:after].to_f))
     end
     if params[:before]
-      @messages = @messages.created_before(Time.at(params[:before].to_i))
+      @messages = @messages.created_before(Time.at(params[:before].to_f))
     end
     @conversation.mark_read! current_user
   end
